Minor fix: avoid possible NPE when handling another Exception.

+++ src/main/java/org/linagora/linshare/core/business/service/impl/DocumentEntryBusinessServiceImpl.java
@@ -541,7 +541,9 @@ public class DocumentEntryBusinessServiceImpl implements DocumentEntryBusinessSe
 			return documentRepository.create(document);
 		} catch (Exception e) {
 			fileSystemDao.removeFileByUUID(uuid);
-			fileSystemDao.removeFileByUUID(uuidThmb);
+			if (uuidThmb != null) {
+				fileSystemDao.removeFileByUUID(uuidThmb);
+			}
 			throw e;
 		}
 	}
